Moved concepts to types/
