+++ include/readout/types/LatencyBufferConcept.hpp
