+++ include/readout/types/RawDataProcessorConcept.hpp
