+++ plugins/FakeCardReader.hpp
@@ -14,7 +14,7 @@
 #define READOUT_PLUGINS_FAKECARDREADER_HPP_
 
 // package
-#include "readout/SourceEmulatorConcept.hpp"
+#include "readout/types/SourceEmulatorConcept.hpp"
 #include "readout/types/ReadoutTypes.hpp"
 #include "readout/utils/ReusableThread.hpp"
 #include "readout/fakecardreader/Structs.hpp"
